feat: updated the improving_loop_times article - Correctly points to CachingHardware, so that the example is not   invalid, and allows users to easily obtain CachingHardware. - Uses files and rustdoc_include to provide fully valid files for   examples, instead of not including imports, while hiding unnecessary   details. - No longer suggests using both Photon and CachingHardware.
